+++ src/main/kotlin/no/nav/bidrag/beregn/bidragsevne/dto/coredto.kt
@@ -41,7 +41,7 @@ data class BostatusPeriodeCore(
 data class BarnIHusstandPeriodeCore(
   val referanse: String,
   val periode: PeriodeCore,
-  val antallBarn: Int
+  val antallBarn: Double
 )
 
 data class SaerfradragPeriodeCore(
